requirements.txt
psycopg2
Flask
Flask-RESTful
Flask-SQLAlchemy
# marshmallow 2.* not yet released, introduces decorators
git+https://github.com/marshmallow-code/marshmallow.git
